refactor(contacts): share ledger sync page props
