+++ apps/ledger-live-desktop/src/mvvm/features/Contacts/screens/Contacts/ContactsView.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { ContactsPage, type ContactsPageWebProps } from "@features/flow-contacts";
+import { ContactsPage, type ContactsPageProps } from "@features/flow-contacts";
 
-export type ContactsViewProps = ContactsPageWebProps;
+export type ContactsViewProps = ContactsPageProps;
 
 export function ContactsView(props: Readonly<ContactsViewProps>) {
   return <ContactsPage {...props} />;
