+++ features/flow/contacts/src/list/index.ts
@@ -3,6 +3,8 @@ export {
   createPopulatedContactsListViewModel,
 } from "./viewModel";
 export type {
+  ContactsLedgerSyncIntroduction,
+  ContactsLedgerSyncStatus,
   ContactsPageLabels,
   ContactsPageProps,
   ContactsListItem,
