+++ features/flow/contacts/src/list/web.ts
@@ -1,2 +1,1 @@
 export { ContactsPage } from "./components/ContactsPage/ContactsPage.web";
-export type { ContactsLedgerSyncStatus, ContactsPageWebProps } from "./ledgerSync/types";
